EvaluationModule/llm-leaderboard/pyproject.toml
[project]
name = "llm-leaderboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "jsonschema>=4.26.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]
